Run Coverage with matrix build
 ### Motivation

Currently, only the execution result of Python 3.11 is included in the coverage measurement range.
However, there are codes that are not executed in Python 3.11, such as workarounds for Python 3.8.
By using matrix build, we can get the execution result of all versions, measure the coverage, and upload it to codecov.