Coverage Correction (2)
When using `from typing_extensions import Protocol` instead of `from typing import Protocol` because Python 3.7 or earlier does not
support it, there is no need to include it in the coverage measurement, so add a `no cover` comment.
By adding the `no cover` comment, you can exclude it from the coverage measurement.